requirements.txt
# VaultKey – Python Desktop Password Manager
# Install with:  pip install -r requirements.txt

# GUI
PySide6>=6.6.0

# HTTP & Firebase REST API
requests>=2.31.0

# Google OAuth – Installed Application flow (desktop)
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0

# Cryptography
cryptography>=42.0.0
bcrypt>=4.1.0

# Environment variables
python-dotenv>=1.0.0
